backend/requirements.txt
# requirements.txt

asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-rest-framework==0.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-stubs==3.15.2
idna==3.10
mypy==1.13.0
mypy-extensions==1.0.0
paho-mqtt==2.1.0
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
